backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
psycopg[binary]>=3.2.0
google-genai>=0.2.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
